+++ src/UglyToad.PdfPig/Content/PageSize.cs
@@ -135,11 +135,11 @@
 
         private struct WidthHeight
         {
-            public decimal Width { get; }
+            public double Width { get; }
 
-            public decimal Height { get; }
+            public double Height { get; }
 
-            public WidthHeight(decimal width, decimal height)
+            public WidthHeight(double width, double height)
             {
                 Width = width;
                 Height = height;
