+++ src/UglyToad.PdfPig/Graphics/Operations/TextPositioning/MoveToNextLine.cs
@@ -32,7 +32,7 @@
         /// <inheritdoc />
         public void Run(IOperationContext operationContext)
         {
-            var tdOperation = new MoveToNextLineWithOffset(0, -1 * operationContext.GetCurrentState().FontState.Leading);
+            var tdOperation = new MoveToNextLineWithOffset(0, -1 * (decimal)operationContext.GetCurrentState().FontState.Leading);
 
             tdOperation.Run(operationContext);
         }
